Many changes. The server can now accept clients and handle the authentication part. Clients too can now connect to a server and authenticate themselves.
For now, the server accepts only one client, and exits right after the client disconnects.
The reason is that libssh doesn't allow yet to poll() on server's socket and all clients
with a single function. There are workarounds, but since an issue is open for this feature and
as it seems that it'll be done soon, I prefer to wait for it and work on other stuff meanwhile.